+++ src/dens.jl
@@ -1,5 +1,4 @@
-struct Dens
-    x
+struct Dens3
     d
     mu
     lims
@@ -10,10 +9,10 @@ end
 Univariate kernel density estimation.
 * `x` : Data (vector) on which the density is estimated.
 * `npoints` : Nb. points estimated within the range of `lims`.
-* `lims` : If `nothing`, this is (minimum(`x`), maximum(`x`)).
+* `lims` : If `nothing`, this is `[minimum(x); maximum(x)]`.
 
 The function uses function `Makie.KernelDensity.kde`.
-After estimation, the estimated are normalized to sum to 1. 
+After estimation, the densities (`d`) are normalized to sum to 1. 
 
 ## Examples
 ```julia
@@ -22,48 +21,50 @@ x = randn(n)
 
 res = dens(x)
 pnames(res)
+res.d     # Density (normalized to sum to 1)
 res.lims  # Range of x
 res.mu    # Mean expected density value for a uniform distribution
-f, ax = lines(res.x, res.d;
+f, ax = lines(res.d.x, res.d.d;
     axis = (xlabel = "x", ylabel = "Density"))
 hlines!(ax, res.mu; color = :grey, linestyle = "-")
 f
 
 xnew = [-4; -1; -3.8; 1; 4]
-d = Jchemo.predict(res, xnew)
-scatter!(ax, xnew, d; color = :red)
+d = Jchemo.predict(res, xnew).d
+scatter!(ax, xnew, d.d; color = :red)
 f
 ```
 """ 
 function dens(x; npoints = 2^8, lims = nothing)
     zx = Float64.(vec(x))
     isnothing(lims) ? lims = (minimum(zx), maximum(zx)) : nothing
-    vfm = Makie.KernelDensity.kde(zx; npoints = npoints,
+    fm = Makie.KernelDensity.kde(zx; npoints = npoints,
         boundary = lims)
-    d = mweight(vfm.density)  # Sum(d) = 1
-    mu = 1 / npoints          # Average expected density value
-    Dens(vfm.x, d, mu, lims)  # d should be >= mu if "dense" area      
+    d = mweight(fm.density)   # Sum(d) = 1
+    mu = 1 / npoints          # Average expected density value; d should be >= mu if "dense" area
+    s = (fm.x .< lims[1]) .| (fm.x .> lims[2])
+    d = DataFrame(x = fm.x, d = d, out = s)
+    Dens3(d, mu, lims)        
 end
 
 """
-    predict(object::Dens, x)
+    predict(object::Dens2, x)
 Compute Y-predictions from a fitted model.
 * `object` : The fitted model.
 * `x` : Data (vector) for which predictions are computed.
 """ 
-function predict(object::Dens, x)
-    x = vec(x)
-    n = length(object.x)
+function predict(object::Dens3, x)
+    x = Float64.(vec(x))
+    n = length(object.d.x)
     m = length(x)
     s = (x .< object.lims[1]) .| (x .> object.lims[2])
+    d = object.d.d
+    zd = zeros(m)
     zs = findall(s .== 0)
-    d = zeros(m)
-    d[zs] .= vec(interpl(reshape(object.d, 1, n), object.x; wlfin = x[zs]))
-    d
+    zd[zs] .= vec(interpl(reshape(d, 1, n), object.d.x; wlfin = x[zs]))
+    d = DataFrame(x = x, d = zd, out = s)
+    (d = d,)
 end
-
-
-